event_log: add peek function
Sometimes it would be beneficial to be able to see log event history
without disrupting the log.

The new API allows to traverse the log in the reverse chronological
order: each new invocation of the API would return the entry preceding
the one returned in the previous invocation, until the log is
exhausted.

Traversing in the reverse order is necessary in case entries are
removed from the log while traversal happens, this way the traversal
stops once the oldest available entry is reached, but the order of the
retrieved entries is not altered.

Note that the new function keeps a single context, so if more than one
thread attempt to traverse the log concurrently the result would be a
corrupted sequence of log events.

A new option is being added to tpm_log console command which would
allow to show the entire log contents in one invocation.

BRANCH=cr50
BUG=b:63760920
TEST=tried invoking tpm_log command in different sequences:

    right after restart:

   > tpm_log peek
   Event 0 at time 0.976 s
   > tpm_log 1 1
   > tpm_log 2 2
   > tpm_log peek
   Event 2 at time 72.325 s
   Event 1 at time 68.452 s
   Event 0 at time 0.976 s
   >

     invoked 'tpm_log 1 1' more than 64 times (log buffer capacity).

   > tpm_log peek
   Event 1 at time 181.136 s
   Event 1 at time 180.994 s
   ...
   Event 1 at time 169.961 s
   Event 1 at time 169.798 s
   >
     64 entries were displayed

Change-Id: I425d7d2623b37880d02d0a4db777e4b3f29dadc4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>