coral: Move board level TCPC init to happen in pd_task_init
This pulls in CL:592716 done in the Eve FW branch.

When the board level TCPC run as an init hook it will frequently
lead to an EC reset when we are trying to recovery a disconnected
battery, potentially even a reboot loop with the most unlucky
timing.
If we instead call it from the pd_task before tcpc_init is called
then the board init hook can stall the pd_task init until the
battery is out of disconnect mode, or giving up after 2 seconds
in case the battery never seems to recover.
This accomplishes two goals: ensure the PD chips are not reset until
the battery is out of disconnect and delay start of the pd_task
(and PD negotiation) until the battery is out of disconnect state.

BUG=b:64375688
BRANCH=None
TEST=manual testing on Coral proto. Remove battery and
connect/disconnect external Type C charger and verify that the system
powers up.

Change-Id: I5b7d27780f5ad815af1641eb988ae695ab2f620e
Signed-off-by: Scott Collyer <scollyer@google.com>