acpi / coral: Add EC mem space field for keyboard backlight presence
Reserve EC address space offset 0x0a for "device features" (as used by
coreboot) and add information about the presence of a keyboard backlight
as the first item on Coral.

The bit is set by default if backlight support is compiled in (to stay
compatible with the current practice where compiled in == hardware
present). If the board defines CONFIG_PWM_KBLIGHT_BOARD_OPTION, it's
expected to implement board_has_kblight() to fill in the bit.

BUG=b:64705535
BRANCH=none
TEST=change the skuid it triggers on manually to see GOOG0002 appear and
disappear accordingly within Chrome OS.

Change-Id: I420aed1ae8d0fb7da477fc72ac55ea75a9f9350b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>