Update patch set 8
Patch Set 8:

(1 comment)

I must agree that this is a pretty lame approach, but let's face it: we are at v7, v1 was sent over a year ago and it fixes a bug I first encountered a year before that. I think having a fix, however not optimal it may be, is the priority here.

I know the current cros toolchain works, but the point is to be able to update it to a newer GCC version, for which this fix is required.

Patch-set: 8
Reviewer: Gerrit User 1133285 <1133285@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>