cr50: Split servo state machine into its own file
This is the last state machine which used common/device_state.c.  But
servo is more complex than that, because it needs to differentiate
state-isn't-known (debouncing) from state-isn't-knowable (Cr50 driving
EC TX), so it's cleaner to split it out the way we did AP and EC state
machines in previous CLs.

BUG=b:35587387
BRANCH=cr50
TEST=TODO (not ready for review yet)

Change-Id: I964bd36c35f52c8ef7b3ea3793b6e0764e93587c
Signed-off-by: Randall Spangler <rspangler@chromium.org>