Update patch set 5
Patch Set 5:

(2 comments)

(I didn't look at the code yet)

I'm concerned about this change on hammer side, as we don't have a FW branch yet, and we currently only update RW image on our dogfooder population.

Given that RO before this CL will not be able to boot these RW after this CL, we can't push it as-is.

Is there a way to disable this change just for hammer?

In the near future, we're going to push RO updates though, and we can imagine some trick to make sure RO is updated before RW (e.g. we can do it as the same time as we switch from dev-key to premp-key signed RO, or increment the major version in ec version string from v1.1 to v1.2 and have hammerd check that the major/minor version string match).

Patch-set: 5
CC: Gerrit User 1123360 <1123360@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>