servo_v4: Add a CCD 'keepalive' console command.
Servo v4 monitors the SBU lines on USB-C connects in order to determine
which direction to set the orientation of the SBU mux.  However, when
there's traffic on the lines, for example when the EC is rebooting, it
can lead to the wrong conclusions and terminate the USB connection.

This commit adds a CCD keepalive console command.

Additonally, when the Type-C cable is unplugged from the DUT for at
least 900ms, the CCD keepalive will be cleared.  This is such that if
the cable is unplugged and then replugged in a different orientation,
the detection will still work.

BUG=b:64903997
BRANCH=servo
TEST=Flash servo_v4; `keepalive enable`.  Run `dut-control
power_state:reset` > 100 times and verify that the USB connection stays
alive.

Change-Id: I5c8f9ab3361d4f52f906161ab5da471a36725a4e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>