common: add means of displaying branch name
There is a need to be able to indicate what branch an image was built
from.

This patch is adding a mechanism to retrieve the branch name from the
file called 'branch_name' and placed in the EC root directory.

When this file is present, its contents are added to generated
build/<board>/ec_version.h as the value of CROS_BRANCH_NAME, when this
definition is present, the 'version' command output includes an extra
line displaying the branch name.

BRANCH=cr50
BUG=b/64698702

TEST=first build a new image without adding the 'branch_name' file,
     observe that the 'version' command output has not changed.

     Then run 'echo xyz > branch_name' in the top level EC directory,
     rebuild the image and examine the version command output, observe
     the branch name included:

   > vers
   Chip:    g cr50 B2-C
   Board:   0
   RO_A:  * 0.0.10/29d77172
   RO_B:    -1.-1.-1/ffffffff
   RW_A:  * 0.0.23/cr50_v1.1.6972-1b98f724a+
   RW_B:    Error
   BID A:   00000000:00000000:00000000 Yes
   BID B:   cccccccc:cccccccc:cccccccc  No
   Branch:  xyz
   Build:   0.0.23/cr50_v1.1.6972-1b98f724a+
            tpm2:v0.0.289-cb2de5a+
            cryptoc:v0.0.8-6283eee
            2017-09-01 15:04:22 vbendeb@eskimo.mtv.corp.google.com

Change-Id: I7a841e7ff0370fc58a36f9f3c0df1f9b9a4c93d5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>