common: add means of displaying the branch name
There is a need to be able to indicate what branch an image was built
from.

This patch is adding a mechanism to retrieve the branch name from a
predefined file, define it in the generated include file, and add it
to the 'version' command output if defined.

BRANCH=cr50
BUG=b/64698702
TEST=manual

Change-Id: I7a841e7ff0370fc58a36f9f3c0df1f9b9a4c93d5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>