eve: Set VCCIO to forced PWM output mode
Set VCCIO configuration to use forced PWM output mode and to
explicitly disable SLP_S0 effects as this is not used so it
should be made explicit since we are setting this register.

Have this run on sysjump as well to ensure that it applies to
systems that get an update without needing an EC reset.

BUG=b:65026806
TEST=update system with flashrom and read back this register to
ensure that it is set to 0x33.

Change-Id: I14543e39ae06911675e643eb153b82e7966df230
Signed-off-by: Duncan Laurie <dlaurie@google.com>