Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> this is a fork of codesigner then?
> what is the trouble of getting actual json parser?
> third_party/rapidjson?

this signer is not used for anything but generating the self signed Cr50 images, and this is the only signer available on the server. (as the real signer requires physical presence at the time, it is not running on the server and is not a part of chroot per se).

So I thought about saving us the trouble of adding new git trees to the chrome os repo.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>