touchpad_elan: Fix the trace number accessing for width/hight property
Base on Elan I2C driver and Elan touchpad spec, the x trace number
of touching object is in the low nibble of byte. And y trace number
is in the high nibble.
This is a minor bug and difficult to find in operation.
I fix the byte accessing to make it right in palm detection

Signed-off-by: KT Liao <kt.liao@emc.com.tw>
BRANCH=poppy
BUG=None
TEST=Test hammer in EVTEST and check MAJOR/MINOR

Change-Id: I83e5e1f224eaf815914e7001234cdc1b1af22660