Update patch set 4
Patch Set 4:

We're about to do a lot of testing to verify that all the rewritten state machines still work as expected.  You prefer to wait to refactor code until *after* the testing?  Why not do it before, so that the testing validates the code which is closest to what we plan on releasing?

If we wait, the objection will then be that this refactoring makes it harder to cherry-pick changes, so we should not do it then either.

And really, when has this project not been in crunch time?

Patch-set: 4
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>