Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This will requires write access to the EC console to get at any port80 codes, which is currently impossible with H1 unless you flash a debug image..
> 
> Maybe we could kick off a deferred function that would print any new entries in the buffer?

Ah yes, that is true! I like the deferred function idea. If intprint is disabled, we can just dump the port80 prints in a deferred function so that they are available for debug later. I will work on this change and push it for review.

Patch-set: 1
Reviewer: Gerrit User 1003000 <1003000@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>