Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> My thoughts at this midnight: 
>   1. We should ignore the IC types stuff (i.e. CONFIG_TOUCHPAD_VIRTUAL_SIZE) for now, as we won't have sample to verify it now and focusing on things can verify could make things solid) In addition, side benefits is that CONFIG_TOUCHPAD_VIRTUAL_SIZE can have a clear path on commit message of evolution.

Please pick up a Soraka, staff base has a Elan TP with larger FW, that's why I updated the configuration in patch set 9. If we push this CL as-is we end up with a somewhat broken/invalid configuration for staff.

>   2. Although we could check-in in parallel, maybe, explicitly making CQ-DEPEND might help the logic for others than the Relation Chain (it is implicit after merged). I'm actually curious if Relation Chain or CQ-DEPEND is better.....

Sure, it's just that the logic is wrong. Right now you have CL:601814 on top of this CL, so CL:601814 won't be committed before this CL anyway. CQ-DEPEND is harmless but useless in this case.

However, this CL can't be committed before CL:601814 (see the build errors from builder above).

These 2 patches (CL:601814 and this) do not depend on anything else, let's push these first.

> Thurs, orders :
>  - Use unified update protocol: virtual address
>  - hammerd can update with that protocol (even we set paths ugly) , so we can at least push necessary updates if something terribly wrong.
>  - Make ec.bin SHAed for touchpad.bin
>  - Adding touchpad.bin and SHAed ec.bin in image at proper place. At this point, we should at least able to update our old-IC hammer in office gracefully.
>  - Works on IC types difference.

I don't think there is much special work to be done for that, at least not on EC side? But maybe I'm missing something.

>  - Avoid unnecessary/incorrect touchpad update (content check before update)
>  - Finish the factory tools that need extract touchpad information. This is quite separate so if tight on factory schedule, we can do this first.

The rest of the steps look reasonable to me.

Patch-set: 11
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>