Update patch set 1
Patch Set 1:

we should probably make this new command more generic so that it can handle the case where we have multiple TCPCs from multiple vendors on the same board.
alternatively, give this command a more device specific name so it won't conflict if we need to introduce similar command(s) for other TCPCs.

Patch-set: 1
CC: Gerrit User 1119124 <1119124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>