Update patch set 1
Patch Set 1:

as an experiment, after fixing the bug, I modified the code to print the prime value after the loop exits. It prints 38873, which is the correct value for the 4096th prime (I added a value to the original array to keep the full number of encoded bytes in the end).

So, getting 38873 is a good indication that calculations are correct, but still
need a better test.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>