Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

> Patch Set 2:
> 
> btw, I found out that when tag is set to a string of a certain length, git sha1 does not fit into the version command output any more, leave alone the traling '+' for dirty tree state marker.
> 
> Maybe the marker should be moved in getversion.sh to become a prefix of the sha1, not a suffix. This way even if sha1 is truncated, the marker is visible. This is how version command output would look (note that the 'Build:' string still shows the full sha1):
> 
> > vers
> Chip:    g cr50 B2-C
> Board:   0
> RO_A:  * 0.0.10/29d77172
> RO_B:    0.0.10/c2a3f8f9
> RW_A:    0.0.23/DBG/cr50_v1.cr50_9308_B.225-e8b
> RW_B:  * 0.0.23/DBG/cr50_v1.cr50_9308_B.225-+e8
> BID A:   00000000:00000000:00000000 Yes
> BID B:   00000000:00000000:00000000 Yes
> Build:   0.0.23/DBG/cr50_v1.cr50_9308_B.225-+e8befe6
>          tpm2:v0.0.288-4119954
>          cryptoc:v0.0.8-6283eee
>          2017-09-11 17:48:20 vbendeb@eskimo.mtv.corp.google.com


Ok with the implementation but
the minor number was really meant to be a number ...
It's diverging from other cros ec builds and the additional information is limited (we have a redundant 'cr50' eg cr50_v1.cr50_9308_B.225 and given as you noticed the space for strings is limited in the EC).
If you feel that the current incrementing number we are using as minor for branching is not giving you enough information, why not using there only the branch number: e.g. 9308   cr50_v1.9308.225

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1