util: move 'dirty' marker to be a prefix, not a suffix
With longer git SHA1s reported by git commands and longer version
strings, the entire SHA1 could be not fitting into the 32 byte version
of the VERSION string. Let's prepend the 'dirty' marker to the string
instead of adding it in the end - otherwise the marker is not visible
in the 32 byte version of the string.

BRANCH=cr50
BUG=b:64698702
TEST=verified output of running getversion.sh:
   $ ./util/getversion.sh | grep  CROS_EC_VERSION32
   #define CROS_EC_VERSION32 "_v1.9308_B.225-+735fba496"

Change-Id: I42684522beaff9e9714206cfaddaf97e6cd644be
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>