Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

Current CL is in fact built upon that CL. Our goal is to use the same script for both sweetberry and in hdctools. However, sweetberry is in ec and depends on ec to build. We plan to have a discussion on this soon.

Patch-set: 1
Reviewer: Gerrit User 1160860 <1160860@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>